0.31
- Makebite folder browser updated
- Makebite now only includes files with an accepted extension
- SnakeBite should now correctly execute GzsTool and receive output

+++ SnakeBite/Classes/ModManager.cs
@@ -356,8 +356,8 @@ namespace SnakeBite
 
         internal static void DeleteGameArchive()
         {
-            File.Delete(GameArchiveXmlPath);
-            Directory.Delete(GameArchiveDir, true);
+            if(File.Exists(GameArchiveXmlPath)) File.Delete(GameArchiveXmlPath);
+            if(Directory.Exists(GameArchiveDir)) Directory.Delete(GameArchiveDir, true);
         }
 
         internal static void ExtractGameArchive()
